requirements.txt
deep-translator
translators
huggingface-hub
numpy
opencv-python
pillow
ultralytics
gradio
transformers
accelerate
qwen-vl-utils
google-genai 